Support for 64-bit games
If a game profile has cpu_architecture="64bit" then the
VIreioDLLInstaller will create symlinks for the 64-bit DLLs rather than
the 32 bit. For 64-bit mode the following caveats are in place:
- No freespace tracker
- Injection is only available by the create symlink approach
- Only The Vanishing of Ethan Carter currently supported for 64-bit

VRBoost 64-bit binary will work with 64bit Vireio.

I moved the logic of saving the game exe to the registry to the DXProxy
loadDLL method, this way even if games don't hook properly, the correct
profile for the game will load (seen issues with Bioshock 2 in this
regard, and means that 64bit support works properly since the hijackDLL
is just a shell at the moment)