DXHR:DC
Think I Have got this as far as I can. Unfortunately the weapon/hands
sprite looks poor as it is 2D and all attempts to fix have failed. Head
tracking (with VRBoost) has a bit of lag.. I think the game has some
in-built mouse smoothing that can't be turned off, and things can get a
bit flickery. Not ideal, but doesn't appear to be a way to fix.