VP Logo
Added an always present Vireio Perception image (same one as the
perception app) in the top left corner of the screen when using Oculus
Rift mode. This should increase "brand awareness" when people are
watching content streamed by other users as it will be clear what
injection driver they are using. It is unobtrusive as it is outside of
the visible area within the rift