Refactor hotkey binding into VPMENU_BindKey
Previously, each menu function would separately check a variable
'hotkeyCatch', check for any pressed keys, and match up the menu
selection with the hotkey to be assigned. This changes it so that when
you indicate you want to bind a key with VPMENU_BindKey, you set a
function onBindKey which takes care of the matching, which allows all
the handling after that to be centralized in VPMENU().

Note that this is the project's very first usage of C++ lambda
expressions, which might cause problems for out of date versions of the
compiler.