Introduce hotkey-expression classes and make all the hardcoded hotkeys use them
Previously, hotkeys were represented in one of two ways. Either there's
a byte, representing a key code, which gets stored and whatnot; or
there's a big boolean expression in code, representing a combination of
shifts and keys and alternative keys and whatnot.

This commit introduces hotkey expressions, which are a tree-structured
representation of combinations of keys; a hotkey expression is something
like "left-shift and F, or left-control and F". Operator overloading is
used to make these easy to write. All the hard-coded hotkeys from
D3DProxyDevice::HandleControls have now been pulled out of their
if-statements into a big block of variables at the top of the file.
(This is still not ideal, but it is an improvement).

The next step will be to convert the *configurable* hotkeys to use this
representation, untangling all the different config-files to make it
work. And the next step after that will be to move all the random
obscure hotkeys into the config, and make them accessible through the
menus.