Refactor strings handling in ProxyHelper
GetTargetExe was previously a very tricky function: it looks like a
getter, but actually it has the side-effect of setting some other (used)
member variables, and also it returns an allocation (which was leaking
memory). This patch makes all the path-getting functions that return
strings, return them as std::string; deletes the tricky member variables
that smuggle return-values through side effects; and reduces log-spam.