Speedup to shader register updating
Profiler results indicated that significant CPU time - in the order of
30% of the portion of the CPU time that's in DxProxy - was being spent
managing the data structures that keeps track of whether shader
registers are dirty, which are std::set<UINT>s. It seems the STL set is
implemented as a tree, which incurs a lot of memory-allocation and
pointer-chasing overhead. Furthermore, register-numbers are reasonably
dense. Replaced the trees with a new vector-based representation, in
class RegisterDirtyStore, for a couple percentage points of framerate
boost.