Use lambda-expressions to manage the currently-open-menu
Rather than having a big enum of all the menus, now, you give
VPMENU_NavigateTo a lambda expression which draws/handles the menu that
you want. This means, first, that there are fewer places to edit to
create a menu; and second, you can give menus parameters, without
needing to create member variables for them (they go in a closure).