Simplify menus with Add*MenuItem functions
Previously, the menu functions would do all input-checking first, then
start drawing. But they could just as easily interleave the two: draw a
menu item, then if it's selected handle its input. And if they're
interleaved this way, then the combination of drawing and handling input
can be factored into a function.

Hence, AddMenuItem, AddButtonMenuIte, AddNavigationMenuItem, and
AddKeybindMenuItem, each of which draws a menu item with DrawMenuItem
and also handles input if it's selected. As with DrawMenuItem's
advancing rectangle, this keeps track of how much it's drawn, so it can
match up the items it's drawing with the selection cursor (and this is
the same as DrawMenuItem's selection cursor).

Most importantly, this gets key-binding down to one line of menu code
per bind to handle all aspects, which makes it (nearly) ready to start
scaling up the number of available keys and the complexity of what you
can do with them.

The diff looks rewrite-ish, but it's mostly transform by rote refactoring.