Change HotkeyCooldown to use #defines instead of number literals
The way hotkey-cooldown is currently handled is very strange. Having
grown out of a hack on the selection Y-position, the specified cooldowns
don't have a linear unit; rather, they're on a logarithmic scale, and
not one that's consistent between framerates. In preparation for making
it sensible, change all calls to use definitions like COOLDOWN_SHORT
rather than numbers like 2.0f.