ProxyConfig default constructor sets all members to defaults
Rather than the LoadConfig function having default settings embedded in
it, the ProxyConfig constructor sets all its members to default, and
when a setting is not mentioned in the config file, LoadConfig leaves it
alone rather than setting the default itself.

This also means that there can be a defaultConfig global variable, which
always has the correct defaults in it.