Redo menu-selection handling
Redo menu-selection handling. Selection now responds instantaneously -
if you press down and then enter, only one frame apart, the "down" is
registered, where before you'd have to wait for the animation to finish.
Also, selection wraps at the top and bottom.

Note that scrolling (as used in the DataGatherer menus) is probably
broken right now, and the scrolling instructions in the
world-scale/convergence menus are as well, since they were weirdly
mingled with the selection cursor.