Add convergence toggle, default disabled
Convergence is a 3D effect which rotates the eyes inward. It is mainly
for 3D monitors; on a head-mounted display, it's wrong. Previously,
there was no way to turn it off - you could only set the convergence
distance far away. On top of that, there was a lot of confusion among
users, who didn't realize that the correct setting was to set the
convergence distance as far away as it would go, to minimize the error.

Convergence is now a toggle, default off. For HMD users, this is as it
should be. 3D monitor users can turn it back on.

There's a bit of help text explaining the HMD/monitor distinction, which
is not currently displayed because MenuBuilder::AddText is a
placeholder.