ViewAdjustment refactoring
Converted ViewAdjustment getters to single lines in the header file, and
grouped left/right matrix pairs together so they can share a comment.
Note that some comments on the getters in ViewAdjustment were deleted by
this; that's intentional, since they duplicated information on the
variables in the header, but had some obvious misinformation ("left"
functions with a comment saying "right", for example).