Correctness fixes for position tracking.
This makes two major fixes to position tracking. First, it changes the
application order of pitch and yaw to positionTransformVec; it was
(pitch*yaw*roll), now it's (yaw*pitch*roll). According to
http://en.wikipedia.org/wiki/Davenport_chained_rotations this is the
correct ordering, and a major artifact went away when I made this
change. (Specifically: it used to be that, if you translate back, look
right, and then pitch up and down, the pitch changes will cause camera
translation.)

The second change is that we capture the game's projection matrix, and
use it to get from projected coordinates to game-worldspace and back.
This eliminates the need to define a position-tracking multiplier on
three separate axes; those multipliers should now all be 1.0.
Unfortunately, there is still a need for a global position tracking
multiplier, for some reason. In Skyrim, I empirically measured it to be
about 50. But this suggests something is still wrong in the
coordinate-space management.

Also note that, while the overall defaults for position tracking
multipliers have been updated, each game has game-specific defaults
which haven't been touched.