Comfort-mode is always on rather than a toggle setting
"Comfort mode" was set up as a toggle, but it makes more sense to just
have it always on; if you don't want to use it, you can now unbind the
hotkeys instead. (In previous versions, this wasn't true; that toggle
was the only way to recover use of the right thumb X-axis if you didn't
want to use it for comfort mode turning). Also, the setting wasn't
saved properly, which was annoying.