Fix a fairly major problem in distortion handling
Aspect ratio correction was being applied incorrectly, causing
distortion to be applied in a coordinate space that was squished on the
Y-axis by 2.0/(eye-buffer aspectRatio) or about 1.125, causing objects
near the top and bottom of the view to be vertically stretched.

I found evidence of this exact mistake outside of Perception, in
https://forums.oculus.com/viewtopic.php?t=2612. I was dubious at first
but after both A/B testing and reasoning it through, I'm confident that
it was wrong before, and that the fix is correct.  This was tricky to
identify, because the artifact it produces (distortion at the edges) is
similar to the artifact that incorrect FOV produces.