Convert to SDK 0.6
This converts the codebase from using SDK 0.4.x to SDK 0.6.x. Note that
you must download the new SDK and put the LibOVR and LibOVRKernel
directories into the source tree, or it will not compile.

This might've broken SDK pose-prediction. I'm not actually sure if that
was working before, but it's even less likely to be working now.