Introduce key-rebinding system
The way this works is, you assign buttons (or button combinations) to
keys, which get sent to the game with SendEvent. The idea is that you
use this to map gamepad buttons to keys on the keyboard, and this lets
you reconcile Perception's hotkeys with games' existing control scheme.
The advantage of this over using XPadder or AutoHotkey is that (a) you
can reconfigure it in-game, which is much easier than quitting and
restarting, and (b) it knows whether the overlay menu is open, and will
turn those keys off until the menu is closed.

You can also set an exclusion with each binding. The idea is that you
use this for setting up combinations with meta-keys; for example, you
might say button X means jump, but RTrigger+X means switch weapon. You
would create a binding for button X, with RTrigger as an exclusion, so
that pressing RTrigger+X doesn't also cause you to jump.