Refactor motion-tracker mouse emulation
This is the first of a two-part refactor. The various overrides of
MotionTracker::updateOrientationPosition show signs of significant
copy-paste damage; some of them use the multipliers
(config->yaw_multiplier etc) and some don't, and they handle wraparound
around 0/360 degrees differently (and not all of them correctly), some
of them check the mouseEmulation option and some don't, and they treat
getOrientationAndPosition as having different conventions for its return
value.

Step one of this refactor is getting the nasty WinAPI SendInput call
wrapped safely in a function, and the INPUT struct it uses out of the
class and into a local variable, but not changing any behavior. Step two
(next commit) will be behavior changes: make everything handle
multipliers and wraparound and the mouseEmulation option correctly, and
probably reduce this to a single non-overridden function.