Heavily refactor motion tracking
Heavily refactored motion tracking. As much as possible is moved out of
the tracker-type-specific files and into the base class. Most of the
stuff in MotionTracker is now private; positions and angles are accessed
through getters. Angle-arithmetic uses two new utility functions,
AngleDifferenceRad/Deg to handle wraparound.

This is important because getting timewarp working is going to require
using the tracker more carefully, referring to positions at multiple
points in time.