TESV - Fog/Smoke Saders
Identified the smoke/fog shaders that cause issue on AMD cards, greatly
reduces the annoying flicker/occlusion artifcact at the expense of
missing smoke/fog effects. Not a disaster, but better than the problem.
Ideally would figure out what the actual issue is, but for now this will
do as a work around
Also identified shaders for clothing...